Refactor Bittorrent client
